pyproject.toml
[project]
name = "BanditLab"
version = "0.1.0"
description = "Framework for multi-armed bandits with support for contextual, GLM, NN, GP-based and delayed methods"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }

authors = [
  { name = "Egor Suraveykin", email = "surikkatik@gmail.com" }
]

dependencies = [
  "numpy>=2.0",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "coba>=8.0"
]

[project.optional-dependencies]
deep = ["torch>=2.0"]
optimization = ["cvxpy>=1.0"]
plots = ["matplotlib>=3.7"]

[tool.setuptools.packages.find]
include = ["mab_framework*"]

[project.scripts]
bandit = "mab_framework.run:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"